feat(profile-creator): implement Profile Creator component and update metadata
